big refactor... probably more to come

+++ src/util/TimeTool.h
@@ -3,9 +3,11 @@
 #include <chrono>
 #include <ctime>
 
+#include "Typedefs.h"
+
 namespace util
 {
-    static inline long long getEpochSeconds()
+    static inline u64 getEpochSeconds()
     {
         // Get the current time as a time_point
         auto now = std::chrono::system_clock::now();
@@ -14,7 +16,7 @@ namespace util
         return std::chrono::duration_cast<std::chrono::seconds>(now.time_since_epoch()).count();
     }
 
-    static inline long long getEpochMilliseconds()
+    static inline u64 getEpochMilliseconds()
     {
         // Get the current time as a time_point
         auto now = std::chrono::system_clock::now();
@@ -23,7 +25,7 @@ namespace util
         return std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
     }
 
-    static inline long long dateToEpochSeconds(uint32_t year, uint32_t month, uint32_t day)
+    static inline u64 dateToEpochSeconds(u32 year, u32 month, u32 day)
     {
         std::tm timeInfo = {};
         timeInfo.tm_year = year - 1900; // Year since 1900
